highlight range between checkin and hovering date
